Refactored matches and recommendation entries into widget classes
